requirements.txt
fastapi==0.114.2
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
httpx==0.27.2
Pillow==10.4.0
mss==9.0.1
pyautogui==0.9.54
sqlite-utils==3.36

